From dev to main (week #4 & 5)

+++ src/main/java/ru/skypro/homework/service/impl/AdServiceImpl.java
@@ -1,10 +1,14 @@
 package ru.skypro.homework.service.impl;
 
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.StringUtils;
 import org.springframework.web.multipart.MultipartFile;
+import ru.skypro.homework.dto.account.Role;
 import ru.skypro.homework.dto.ads.Ads;
 import ru.skypro.homework.dto.ads.CreateAds;
 import ru.skypro.homework.dto.ads.FullAds;
@@ -24,9 +28,11 @@ import ru.skypro.homework.service.ResponseWrapperCommentMapper;
 
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.List;
 
+@Slf4j
 @Service
 public class AdServiceImpl implements AdService {
     private final CommentRepository commentRepository;
@@ -34,6 +40,7 @@ public class AdServiceImpl implements AdService {
     private final UserRepository userRepository;
     private final AdRepository adRepository;
     private final AdMapper adMapper;
+    private final UserDetails userDetails;
 
     @Value("${ads.image.dir.path}")
     private String adsImageDir;
@@ -42,12 +49,14 @@ public class AdServiceImpl implements AdService {
                          ResponseWrapperCommentMapper responseWrapperCommentMapper,
                          UserRepository userRepository,
                          AdRepository adRepository,
-                         AdMapper adMapper) {
+                         AdMapper adMapper,
+                         UserDetails userDetails) {
         this.commentRepository = commentRepository;
         this.responseWrapperCommentMapper = responseWrapperCommentMapper;
         this.userRepository = userRepository;
         this.adRepository = adRepository;
         this.adMapper = adMapper;
+        this.userDetails = userDetails;
     }
 
     @Override
@@ -60,12 +69,12 @@ public class AdServiceImpl implements AdService {
 
     @Override
     @Transactional
-    public Comment addComment(Integer id, CreateComment createComment, String userName) {
+    public Comment addComment(Integer id, CreateComment createComment) {
         AdEntity adEntity = adRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("Объявление не найдено"));
         CommentEntity commentEntity = responseWrapperCommentMapper.toCommentEntity(createComment, new CommentEntity());
-        UserEntity userEntity = userRepository.findByEmail(userName)
-                .orElseThrow(() -> new IllegalArgumentException("Пользователь не найден"));
+        UserEntity userEntity = userRepository.findByEmail(userDetails.getUsername())
+                .orElseThrow(() -> new UsernameNotFoundException("Пользователь не найден"));
         commentEntity.setUserEntity(userEntity);
         commentEntity.setAdEntity(adEntity);
         return responseWrapperCommentMapper.toCommentDto(commentRepository.save(commentEntity));
@@ -73,32 +82,50 @@ public class AdServiceImpl implements AdService {
 
     @Override
     @Transactional
-    public void deleteComment(Integer adId, Integer commentId) {
-        commentRepository.delete(commentRepository
-                .findByIdAndAdEntity_Id(commentId, adId));
+    public boolean deleteComment(Integer adId, Integer commentId) {
+        String userName = userDetails.getUsername();
+        UserEntity userEntity = userRepository.findByEmail(userName)
+                .orElseThrow(() -> new UsernameNotFoundException("Пользователь не найден"));
+        CommentEntity commentEntity = commentRepository.findByIdAndAdEntity_Id(commentId, adId)
+                .orElseThrow(() -> new IllegalArgumentException("Комментарий не найден"));
+        if (userCanChangeComment(userEntity, commentEntity)) {
+            commentRepository.delete(commentRepository
+                    .findByIdAndAdEntity_Id(commentId, adId)
+                    .orElseThrow(() -> new IllegalArgumentException("Комментарий не найден"))
+            );
+            return true;
+        }
+        return false;
     }
 
     @Override
     @Transactional
     public Comment updateComment(Integer adId, Integer commentId, CreateComment createComment) {
-        CommentEntity commentEntity = commentRepository.findByIdAndAdEntity_Id(commentId, adId);
-        return responseWrapperCommentMapper.toCommentDto(
-                commentRepository.save(
-                        responseWrapperCommentMapper.toCommentEntity(
-                                createComment, commentEntity)
-                )
-        );
+        String userName = userDetails.getUsername();
+        UserEntity userEntity = userRepository.findByEmail(userName)
+                .orElseThrow(() -> new UsernameNotFoundException("Пользователь не найден"));
+        CommentEntity commentEntity = commentRepository.findByIdAndAdEntity_Id(commentId, adId)
+                .orElseThrow(() -> new IllegalArgumentException("Комментарий не найден"));
+        if (userCanChangeComment(userEntity, commentEntity)) {
+            return responseWrapperCommentMapper.toCommentDto(
+                    commentRepository.save(
+                            responseWrapperCommentMapper.toCommentEntity(
+                                    createComment, commentEntity)
+                    )
+            );
+        }
+        return null;
     }
 
     @Override
     @Transactional
-    public Ads addAdvertising(CreateAds createAds, MultipartFile image, String userName) throws IOException {
-        UserEntity user = userRepository.findByEmail(userName)
-                .orElseThrow(() -> new IllegalArgumentException("Пользователь не найден"));
+    public Ads addAdvertising(CreateAds createAds, MultipartFile image) throws IOException {
+        UserEntity user = userRepository.findByEmail(userDetails.getUsername())
+                .orElseThrow(() -> new UsernameNotFoundException("Пользователь не найден"));
         AdEntity adEntity = adRepository.save(adMapper.toAdEntity(createAds, new AdEntity()));
         Path filePath = createPath(image, adEntity);
         adEntity.setUserEntity(user);
-        adEntity.setImagePath(filePath.getParent().toString());
+        adEntity.setImagePath(filePath.toAbsolutePath().toString());
         adEntity.setImageMediaType(image.getContentType());
         adEntity.setImageFileSize(image.getSize());
 
@@ -122,21 +149,27 @@ public class AdServiceImpl implements AdService {
 
     @Override
     @Transactional(readOnly = true)
-    public ResponseWrapperAds getAllMyAdvertising(String userName) {
-        List<AdEntity> entityList = adRepository.findAllByUserEntityEmail(userName);
+    public ResponseWrapperAds getAllMyAdvertising() {
+        List<AdEntity> entityList = adRepository.findAllByUserEntityEmail(userDetails.getUsername());
         return adMapper.toResponseWrapperAds(entityList);
     }
 
     @Override
     @Transactional
     public Ads updateAdvertising(int id, CreateAds createAds) {
+        String userName = userDetails.getUsername();
+        UserEntity userEntity = userRepository.findByEmail(userName)
+                .orElseThrow(() -> new UsernameNotFoundException("Пользователь не найден"));
         AdEntity adEntity = adRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("Объявление не найдено"));
-        return adMapper.toAds(
-                adRepository.save(
-                        adMapper.toAdEntity(createAds, adEntity)
-                )
-        );
+        if (userCanChangeAdvertising(userEntity, adEntity)) {
+            return adMapper.toAds(
+                    adRepository.save(
+                            adMapper.toAdEntity(createAds, adEntity)
+                    )
+            );
+        }
+        return null;
     }
 
     @Override
@@ -144,8 +177,11 @@ public class AdServiceImpl implements AdService {
     public boolean updateAdvertisingImage(int id, MultipartFile image) throws IOException {
         AdEntity adEntity = adRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("Объявление не найдено"));
+        if (adEntity.getImagePath() != null) {
+            Files.deleteIfExists(Path.of(adEntity.getImagePath()));
+        }
         Path filePath = createPath(image, adEntity);
-        adEntity.setImagePath(filePath.getParent().toString());
+        adEntity.setImagePath(filePath.toAbsolutePath().toString());
         adEntity.setImageMediaType(image.getContentType());
         adEntity.setImageFileSize(image.getSize());
         adRepository.save(adEntity);
@@ -154,10 +190,19 @@ public class AdServiceImpl implements AdService {
 
     @Override
     @Transactional
-    public boolean deleteAdvertising(int id) {
-        commentRepository.deleteAllByAdEntity_Id(id);
-        adRepository.deleteById(id);
-        return true;
+    public boolean deleteAdvertising(int id) throws IOException {
+        String userName = userDetails.getUsername();
+        UserEntity userEntity = userRepository.findByEmail(userName)
+                .orElseThrow(() -> new UsernameNotFoundException("Пользователь не найден"));
+        AdEntity adEntity = adRepository.findById(id)
+                .orElseThrow(() -> new IllegalArgumentException("Объявление не найдено"));
+        if (userCanChangeAdvertising(userEntity, adEntity)) {
+            commentRepository.deleteAllByAdEntity_Id(id);
+            adRepository.deleteById(id);
+            Files.deleteIfExists(Path.of(adEntity.getImagePath()));
+            return true;
+        }
+        return false;
     }
 
     @Override
@@ -176,6 +221,7 @@ public class AdServiceImpl implements AdService {
                 adEntity.getImagePath(),
                 adEntity.getImageMediaType(),
                 adEntity.getImageFileSize());
+        log.info("The method was called to download ads image with title " + adEntity.getTitle());
     }
 
     private Path createPath(MultipartFile image, AdEntity adEntity) throws IOException {
@@ -183,5 +229,13 @@ public class AdServiceImpl implements AdService {
                 + StringUtils.getFilenameExtension(image.getOriginalFilename()));
         AccountServiceImpl.uploadImage(image, filePath);
         return filePath;
+    }
+
+    private boolean userCanChangeAdvertising(UserEntity user, AdEntity ad) {
+        return user.getRole() == Role.ADMIN || ad.getUserEntity().equals(user);
+    }
+
+    private boolean userCanChangeComment(UserEntity user, CommentEntity comment) {
+        return user.getRole() == Role.ADMIN || comment.getUserEntity().equals(user);
     }
 }
